requirements.txt
apache-airflow
pandas
psycopg2-binary
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
oauth2client<4.0.0
minio
boto3
